Implemented Expanded Job run component
